Moved the generated code to the "Federation" folder.
The generated CMake file in the root of the Federation folder sets up a solution for the Federation code. Visual Studio packages all of the HLA code into a .dll file, and we use that .dll file as an include in another project within the solution.

Everything should be generated automatically if you run cmake --build in the Federation directory. If that doesn't work then download the CMake GUI and do it that way.